Update 1.9.5_2 (#390) (#391)
#390
