+++ config/example_options.ini
@@ -6,16 +6,27 @@
 ; The bot must be restarted for edits to take effect, but a reload command will be created in the future.
 ; Currently the bot does not overwrite any settings, but this may change in a future update.
 
+
+; HOW TO GET VARIOUS IDS:
+; http://i.imgur.com/GhKpBMQ.gif
+; Enable developer mode (options, settings, appearance), right click the object you want the id of, and click Copy ID
+; This works for basically everything you would want the id of (channels and users).  For roles you have to right click a role mention.
+
+
 [Credentials]
-; The bot needs its own account to run on (not yours)
-Email = bot_discord@email
-Password = bot_discord_password
-; If you have a bot account, use this line instead.  Comment or remove the Email/Password options and uncomment the Token line.
-;Token = bot_token
+; Put your token here.  Not "secret".  The secret is not the token.
+Token = bot_token
+; If you want to use a normal user account instead of a bot account, use these lines instead.
+; Comment the token line and uncomment the Email/Password lines.
+;Email = bot_discord@email
+;Password = bot_discord_password
 
 [Permissions]
-; This need to be your id, not the bot's id.  The bot does not own the bot.  You do.
-; If you don't know how to get this, join the help server (link in the readme) *with the owner's account* and type !id in chat
+; This number should be your id.  It gives you full permissions.  You do not put the bot's id here.  That's silly.
+; If you don't know how to get this, scroll up a bit and read the part that says "HOW TO GET VARIOUS IDS"
+; If you can't do that for some reason, join the help server (invite in the readme) and type this in chat: !id
+; If you still don't understand, watch this https://streamable.com/4w8e and may your respective deity have mercy on your soul.
+; I don't want any more "how do I get the OwnerID" questions.
 OwnerID = 000000000000000000
 
 [Chat]
