+++ requirements.txt
@@ -1,2 +1,5 @@
-https://github.com/Rapptz/discord.py/archive/async.zip
-youtube_dl >= 2016.4.24
+discord.py[voice]
+youtube_dl
+pip
+cffi --only-binary cffi; sys_platform == 'win32'
+cffi; sys_platform != 'win32'
